Change method in 3rd task

+++ js/task-3.js
@@ -1,5 +1,5 @@
 function getElementWidth(content, padding, border) {
-    const width = ((Number.parseInt(content)) + 2 * (Number.parseInt(padding)) + 2 * (Number.parseFloat(border))); 
+    const width = ((Number.parseFloat(content)) + 2 * (Number.parseFloat(padding)) + 2 * (Number.parseFloat(border))); 
     return width;
 }
